Added details admin pages

+++ backend/src/detail/detail.service.ts
@@ -1,13 +1,19 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { Detail, Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { UsersService } from 'src/users/users.service';
 
 @Injectable()
 export class DetailService {
-    constructor(private prisma: PrismaService) { }
-    
-    async findOne(where: Prisma.DetailWhereUniqueInput): Promise<Detail | null> {
-        return await this.prisma.detail.findUnique({where});
+    constructor(
+        private prisma: PrismaService,
+        private usersService: UsersService
+    ) { }
+
+    async findOne(
+        where: Prisma.DetailWhereUniqueInput
+    ): Promise<Detail | null> {
+        return await this.prisma.detail.findUnique({ where });
     }
 
     async findMany(params: {
@@ -27,11 +33,26 @@ export class DetailService {
     async update(params: {
         where: Prisma.DetailWhereUniqueInput;
         data: Prisma.DetailUpdateInput;
-    }): Promise<Detail> {
+    }, username: string): Promise<Detail> {
+        const user = await this.usersService.findOneByUsername(username);
+        const detail = await this.prisma.detail.findUnique({
+            where: params.where
+        });
+        if (detail.seller_id !== user?.id) {
+            throw new ForbiddenException();
+        }
         return await this.prisma.detail.update(params);
     }
 
-    async delete(where: Prisma.DetailWhereUniqueInput): Promise<Detail> {
+    async delete(
+        where: Prisma.DetailWhereUniqueInput,
+        username: string
+    ): Promise<Detail> {
+        const user = await this.usersService.findOneByUsername(username);
+        const detail = await this.prisma.detail.findUnique({ where });
+        if (detail.seller_id !== user?.id) {
+            throw new ForbiddenException();
+        }
         return await this.prisma.detail.delete({ where });
     }
 }
